- Implement custom hook 'useDynamicTransition'
