+++ 15dkatz-hooks-course/src/hooks.js
@@ -19,3 +19,17 @@ export const useFetch = (url, initialValue) => {
 
   return result;
 };
+
+export const useDynamicTransition = ({ delay, increment, picsLength }) => {
+  const [index, setIndex] = useState(0);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setIndex((storedIndex) => (storedIndex + increment) % picsLength);
+    }, delay);
+
+    return () => clearInterval(interval);
+  }, [delay, increment]);
+
+  return index;
+};
